pyproject.toml
[project]
name = "affective-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "matplotlib>=3.10.8",
    "pyyaml>=6.0",
]
